配置调整

+++ drama/src/main/java/com/cute/play/drama/entity/BaseEntity.java
@@ -18,5 +18,5 @@ class BaseEntity {
     private String modifier;
     private Date modified;
     @TableLogic
-    private Integer deleted;
+    private Byte deleted;
 }
